requirements.txt
celery==4.3.0
django-crispy-forms==1.7.2
Django==1.11.25
django-celery==3.3.1
django-celery-beat==1.5.0
django_celery_results==1.1.2
django-filter==1.0.4
django-markdown==0.8.4
django-pagedown==0.1.3
django-widget-tweaks==1.4.2
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Pillow==6.2.0
psycopg2==2.8.3
requests==2.22.0
beautifulsoup4==4.8.1